move property_to_s() to a ValueHelper
needs to be accessed by the mailer
